requirements.txt
asgiref==3.8.1
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.4.0
django-oidc-provider==0.8.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
future==1.0.0
gunicorn==22.0.0
idna==3.7
mysqlclient==2.2.4
packaging==24.1
pillow==10.4.0
pycryptodomex==3.20.0
pyjwkest==1.4.2
PyJWT==2.8.0
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.2
